+++ skiko/src/jvmMain/cpp/linux/drawlayer.cc
@@ -110,7 +110,7 @@ LayerHandler *findByObject(JNIEnv *env, jobject object)
 
 extern "C"
 {
-    JNIEXPORT void JNICALL Java_org_jetbrains_skiko_layer_HardwareLayer_updateLayer(JNIEnv *env, jobject canvas)
+    JNIEXPORT void JNICALL Java_org_jetbrains_skiko_HardwareLayer_updateLayer(JNIEnv *env, jobject canvas)
     {
         if (layerStorage != NULL)
         {
@@ -173,7 +173,7 @@ extern "C"
         awt.FreeDrawingSurface(ds);
     }
 
-    JNIEXPORT void JNICALL Java_org_jetbrains_skiko_layer_HardwareLayer_redrawLayer(JNIEnv *env, jobject canvas)
+    JNIEXPORT void JNICALL Java_org_jetbrains_skiko_HardwareLayer_redrawLayer(JNIEnv *env, jobject canvas)
     {
         LayerHandler *layer = findByObject(env, canvas);
         if (layer != NULL)
@@ -182,7 +182,7 @@ extern "C"
         }
     }
 
-    JNIEXPORT void JNICALL Java_org_jetbrains_skiko_layer_HardwareLayer_disposeLayer(JNIEnv *env, jobject canvas)
+    JNIEXPORT void JNICALL Java_org_jetbrains_skiko_HardwareLayer_disposeLayer(JNIEnv *env, jobject canvas)
     {
         LayerHandler *layer = findByObject(env, canvas);
         if (layer != NULL)
@@ -193,7 +193,7 @@ extern "C"
         }
     }
 
-    JNIEXPORT jlong JNICALL Java_org_jetbrains_skiko_layer_HardwareLayer_getWindowHandle(JNIEnv *env, jobject canvas)
+    JNIEXPORT jlong JNICALL Java_org_jetbrains_skiko_HardwareLayer_getWindowHandle(JNIEnv *env, jobject canvas)
     {
         JAWT awt;
         JAWT_DrawingSurface *ds = NULL;
