+++ skiko/src/jvmMain/kotlin/org/jetbrains/skiko/HardwareLayer.kt
@@ -1,15 +1,9 @@
-package org.jetbrains.skiko.layer
+package org.jetbrains.skiko
 
 import java.awt.Graphics
 import java.awt.Canvas
 import javax.swing.SwingUtilities.convertPoint
 import javax.swing.SwingUtilities.getRootPane
-import org.jetbrains.skiko.Library
-import org.jetbrains.skiko.hostOs
-import org.jetbrains.skiko.OS
-import org.jetbrains.skiko.properties.MacOSProperties
-import org.jetbrains.skiko.properties.LinuxProperties
-import org.jetbrains.skiko.properties.WindowsProperties
 
 abstract class HardwareLayer : Canvas(), Drawable {
