--- skiko/src/jvmMain/kotlin/org/jetbrains/skiko/render/SkiaRenderer.kt
@@ -1,10 +1,0 @@
-package org.jetbrains.skiko.render
-
-import org.jetbrains.skija.Canvas
-
-interface SkiaRenderer {
-    fun onInit()
-    fun onRender(canvas: Canvas, width: Int, height: Int)
-    fun onReshape(width: Int, height: Int)
-    fun onDispose()
-}
